backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
requests==2.31.0
statsmodels==0.14.1
google-genai==0.6.0
